requirements.txt
Flask==2.3.2
Flask-WTF==1.1.1
Flask-Limiter==2.7.0
Flask-Talisman==1.0.0
Flask-SQLAlchemy==3.0.3
argon2-cffi==21.3.0
WTForms==3.0.1